feat: add getAvailableQuantity for ingredients

+++ src/main/java/com/restaurant/db/migration/restaurant.sql
@@ -38,13 +38,13 @@ CREATE TABLE IF NOT EXISTS dish_ingredient (
 
 CREATE TABLE IF NOT EXISTS stock_in (
     ingredient_id bigint REFERENCES ingredient(ingredient_id),
-    quantity bigint not null,
+    quantity NUMERIC(10,2) not null,
     last_modified timestamp not null
 );
 
 CREATE TABLE IF NOT EXISTS stock_out (
     ingredient_id bigint REFERENCES ingredient(ingredient_id),
-    quantity bigint not null,
+    quantity NUMERIC(10,2) not null,
     last_modified timestamp not null
 );
